pyproject.toml
[tool.poetry]
name = "insight-backend"
version = "0.1.0"
description = ""
authors = ["abundance <abundanceanyanwu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
alembic = "^1.13.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
alembic-autogen-check = "^1.1.1"
ormar = "^0.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
